+++ TeamBST/java/com/itwillbs/service/MemberService.java
@@ -10,5 +10,8 @@ public interface MemberService {
 	
   //로그인 체크
 	public UserDTO memberLogin(UserDTO loginDto);
+
+	//아이디 중복체크
+	public int idCheck(String user_id) throws Exception;
 	
 }
