+++ TeamBST/src/main/resources/mappers/memberMapper.xml
@@ -19,4 +19,18 @@
 	where user_id=#{user_id} and user_pwd=#{user_pwd}
 </select>
 
+
+<!-- 아이디 중복체크 -->
+<select id="idCheck" parameterType="String" resultType="int">
+	select count(*)from user where user_id=#{user_id}
+</select>
+
+<!-- 닉네임 중복체크 -->
+<select id="nickCheck" parameterType="String" resultType="int">
+	select count(*)from user where user_nick=#{user_nick}
+</select>
+<!-- 이메일 중복체크 -->
+<select id="emailCheck" parameterType="String" resultType="int">
+	select count(*)from user where user_email=#{user_email}
+</select>
 </mapper>
