+++ TeamBST/src/main/webapp/META-INF/maven/com.itwillbs/controller/pom.properties
@@ -1,6 +1,6 @@
 #Generated by Maven Integration for Eclipse
-#Thu Jun 13 01:57:32 KST 2024
-m2e.projectLocation=C\:\\Users\\hong5\\git\\TeamBST\\TeamBST
+#Wed Jun 19 09:11:38 KST 2024
+m2e.projectLocation=C\:\\Users\\ITWILL\\git\\TeamBST_real\\TeamBST
 m2e.projectName=TeamBST
 groupId=com.itwillbs
 artifactId=controller
