+++ TeamBST/src/main/webapp/META-INF/maven/com.itwillbs/controller/pom.xml
@@ -174,6 +174,14 @@
 	       <artifactId>HikariCP</artifactId>
            <version>4.0.3</version>
         </dependency>
+        
+        <!-- Jackson 라이브러리 -->
+   	   	<dependency>
+	       <groupId>com.fasterxml.jackson.core</groupId>
+	       <artifactId>jackson-databind</artifactId>
+	       <version>2.9.2</version>
+   	   	</dependency>
+    	<!-- Jackson 라이브러리 -->
 		
 		<!-- 외부 라이브러리 추가 -->        
 	</dependencies>
