+++ TeamBST/src/main/webapp/WEB-INF/views/login/idCheck_proc.jsp
@@ -1,0 +1,12 @@
+<%@ page language="java" contentType="text/html; charset=UTF-8"
+    pageEncoding="UTF-8"%>
+<!DOCTYPE html>
+<html>
+<head>
+<meta charset="UTF-8">
+<title>Insert title here</title>
+</head>
+<body>
+<h1>ㅇㅇㅇ</h1>
+</body>
+</html>
